requirements.txt
cryptography>=41.0.0
discord.py>=2.3.0
